requirements.txt
langgraph
langchain
langchain-community
langchain-openai
chromadb
streamlit
pandas
openpyxl
pypdf
tiktoken
faiss-cpu
numpy
requests
python-dotenv
openai
sentence-transformers
pytest
